[Fix] 하드코딩된 데이터 뷰모델에 연결

+++ ToGather/ToGather/Models/Saving.swift
@@ -15,7 +15,7 @@ struct Saving {
 
     var savingDayOfTheWeek: String
 
-    var savingAmountOfWeek: Int {Int(goalProduct.productPrice / Double(goalWeeks))}
+    var savingAmountOfWeek: Double {goalProduct.productPrice / Double(goalWeeks)}
 
     var weekInfo: [ThisWeek]
